ref(ui-components) Move components to base/ui (#11912)
