+++ react/features/base/react/constants.ts
@@ -3,20 +3,3 @@
  * everything, such as modal-type of components, or dialogs.
  */
 export const OVERLAY_Z_INDEX = 1000;
-
-/**
- * The types of the buttons.
- */
-export enum BUTTON_TYPES {
-    DESTRUCTIVE = 'destructive',
-    PRIMARY = 'primary',
-    SECONDARY = 'secondary',
-    TERTIARY = 'tertiary'
-}
-
-/**
- * The modes of the buttons.
- */
-export const BUTTON_MODES = {
-    CONTAINED: 'contained'
-};
