+++ react/features/base/react/types.ts
@@ -1,12 +1,3 @@
-import { ButtonProps } from '../components/common/types';
-
-export interface IButtonProps extends ButtonProps {
-    color?: string;
-    labelStyle?: Object|undefined;
-    onPress?: Function;
-    style?: Object|undefined;
-}
-
 export interface IconButtonProps {
     accessibilityLabel?: string;
     color?: string;
