+++ react/features/base/ui/components/native/Input.tsx
@@ -10,12 +10,12 @@ import {
     ViewStyle
 } from 'react-native';
 
-import { InputProps } from '../../../components/common/types';
 import Icon from '../../../icons/components/Icon';
 import { IconCloseCircle } from '../../../icons/svg';
 // eslint-disable-next-line lines-around-comment
 // @ts-ignore
 import BaseTheme from '../../../ui/components/BaseTheme.native';
+import { InputProps } from '../types';
 
 import styles from './inputStyles';
