+++ react/features/participants-pane/components/breakout-rooms/components/native/AddBreakoutRoomButton.js
@@ -3,8 +3,8 @@
 import React, { useCallback } from 'react';
 import { useDispatch } from 'react-redux';
 
-import Button from '../../../../../base/react/components/native/Button';
-import { BUTTON_TYPES } from '../../../../../base/react/constants';
+import Button from '../../../../../base/ui/components/native/Button';
+import { BUTTON_TYPES } from '../../../../../base/ui/constants';
 import { createBreakoutRoom } from '../../../../../breakout-rooms/actions';
 
 import styles from './styles';
