+++ react/features/participants-pane/components/native/LobbyParticipantItem.js
@@ -4,8 +4,8 @@ import React, { useCallback } from 'react';
 import { useDispatch } from 'react-redux';
 
 import { hasRaisedHand } from '../../../base/participants';
-import Button from '../../../base/react/components/native/Button';
-import { BUTTON_TYPES } from '../../../base/react/constants';
+import Button from '../../../base/ui/components/native/Button';
+import { BUTTON_TYPES } from '../../../base/ui/constants';
 import { approveKnockingParticipant } from '../../../lobby/actions.native';
 import { showContextMenuReject } from '../../actions.native';
 import { MEDIA_STATE } from '../../constants';
