+++ react/features/participants-pane/components/web/InviteButton.tsx
@@ -5,9 +5,9 @@ import { useDispatch } from 'react-redux';
 
 // @ts-ignore
 import { createToolbarEvent, sendAnalytics } from '../../../analytics';
-import Button from '../../../base/components/common/Button';
 import { IconInviteMore } from '../../../base/icons/svg/index';
-import { BUTTON_TYPES } from '../../../base/react/constants';
+import Button from '../../../base/ui/components/web/Button';
+import { BUTTON_TYPES } from '../../../base/ui/constants';
 // @ts-ignore
 import { beginAddPeople } from '../../../invite';
