+++ react/features/recording/components/Recording/native/HighlightDialog.js
@@ -4,8 +4,8 @@ import { Text, View } from 'react-native';
 import { useDispatch, batch } from 'react-redux';
 
 import { BottomSheet, hideSheet } from '../../../../base/dialog';
-import Button from '../../../../base/react/components/native/Button';
-import { BUTTON_TYPES } from '../../../../base/react/constants';
+import Button from '../../../../base/ui/components/native/Button';
+import { BUTTON_TYPES } from '../../../../base/ui/constants';
 import { highlightMeetingMoment } from '../../../actions.any';
 import styles from '../styles.native';
